+++ tests/test_coding/test_transforms.py
@@ -6,6 +6,9 @@ from satella.coding.transforms import stringify, split_shuffle_and_join, one_tup
 
 class TestTransforms(unittest.TestCase):
 
+    def test_stringify_bug(self):
+        self.assertEqual(stringify('test'), 'test')
+
     def test_clip(self):
         self.assertEqual(clip(5, 10, 15), 10)
         self.assertEqual(clip(25, 10, 15), 15)