+++ CHANGELOG.md
@@ -1,3 +1,5 @@
 # v2.12.1
 
+* bugfix release: fixed a bug with `stringify`
+